+++ docs/latest/integrations/introduction.mdx
@@ -400,6 +400,22 @@ Start integrating and monitoring your applications with OpenLIT effortlessly. Ch
     >
   </Card>
   <Card
+      title="Google AI Studio"
+      href="/latest/integrations/google-ai-studio"
+      icon={
+        <svg version="1.1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 205 205">
+          <path d="M0 0 C8.04946485 6.85473908 11.51216683 14.78808209 13.04296875 25.04296875 C13.72284084 35.15606611 9.7120795 45.22637136 3.10546875 52.79296875 C-4.50478023 59.69884669 -13.19767256 61.98589369 -22.89453125 64.35546875 C-34.66881203 67.26248127 -44.54223956 70.68764928 -53.89453125 78.79296875 C-53.64733504 82.41851319 -53.47317708 84.2474189 -50.77734375 86.7890625 C-47.62185567 88.98251153 -44.4532783 90.51488513 -40.95703125 92.10546875 C-33.36614213 95.75877472 -28.99803821 99.34129857 -26.01953125 107.29296875 C-23.9433434 114.60905926 -24.61486873 120.97595594 -27.89453125 127.79296875 C-32.12953559 134.45133668 -37.27163058 137.88424244 -44.89453125 139.79296875 C-51.59470193 140.85862661 -57.6032759 139.17132116 -63.03515625 135.24609375 C-74.89986413 125.31489382 -81.03911654 108.34091202 -82.6953125 93.24169922 C-85.0510931 64.28522937 -80.12287701 37.80526376 -61.515625 14.859375 C-45.95268612 -3.09036828 -21.6056458 -15.48723283 0 0 Z " fill="#0D5ED0" transform="translate(190.89453125,46.20703125)"/>
+          <path d="M0 0 C15.07697101 12.23417513 24.16302983 28.76270314 27.41943359 47.80615234 C27.54962891 48.53189453 27.67982422 49.25763672 27.81396484 50.00537109 C30.60467234 68.34962173 26.42543136 86.9709579 16.29052734 102.38037109 C13.04453102 106.58864608 9.5020167 109.11195798 4.41943359 110.80615234 C-1.5633288 111.39269768 -5.68297763 110.11734823 -10.58056641 106.80615234 C-15.52138184 101.48527419 -15.9568107 97.88536322 -15.83837891 90.77099609 C-15.47236842 86.56187554 -14.51517635 84.87725578 -11.58056641 81.80615234 C-9.75850995 80.31311671 -7.92384837 78.83540275 -6.08056641 77.36865234 C-1.56726272 73.76524032 1.35201026 71.2208325 3.41943359 65.80615234 C3.69759141 63.04409402 3.61509714 60.59435783 3.41943359 57.80615234 C-3.75922154 53.8180106 -10.187634 52.21543176 -18.35400391 51.45458984 C-29.26661583 50.43494061 -38.78584087 48.24671604 -46.64306641 39.93115234 C-53.04862723 31.60392327 -53.73413155 23.19922494 -52.58056641 12.80615234 C-50.28248956 5.10504149 -45.34440866 -1.4669013 -38.34228516 -5.47509766 C-25.16962991 -11.67399424 -11.46370004 -8.65886518 0 0 Z " fill="#0D5FD0" transform="translate(74.58056640625,66.19384765625)"/>
+          <path d="M0 0 C0.2475 0.556875 0.495 1.11375 0.75 1.6875 C4.07335006 7.83569761 7.70469272 10.13976727 13.875 12.875 C21.86387597 16.66522519 26.67147691 20.00994553 29.875 28.5 C31.95118785 35.81609051 31.27966252 42.18298719 28 49 C23.76499566 55.65836793 18.62290067 59.09127369 11 61 C4.29982932 62.06565786 -1.70874465 60.37835241 -7.140625 56.453125 C-13.69414 50.96759023 -18.4483245 44.21858674 -21 36 C-20.67 35.01 -20.34 34.02 -20 33 C-19.8453125 33.9590625 -19.8453125 33.9590625 -19.6875 34.9375 C-19.3471875 35.9584375 -19.3471875 35.9584375 -19 37 C-18.01 37.33 -17.02 37.66 -16 38 C-15.67 39.98 -15.34 41.96 -15 44 C-14.99742188 42.88625 -14.99484375 41.7725 -14.9921875 40.625 C-14.69527161 24.7004696 -11.19120715 11.81609689 0 0 Z " fill="#8EAEFE" transform="translate(135,125)"/>
+          <path d="M0 0 C9.53843272 -0.37243338 20.25997378 -0.31915184 28 6 C37.11156809 14.87490398 41.7292091 24.92323358 45 37 C45.66 37 46.32 37 47 37 C46.18682587 44.79485489 41.70346121 51.53759248 36.01953125 56.8515625 C31.05157575 60.47425073 27.08524026 61.49879019 21 61 C14.94098975 58.92004126 10.91301855 55.8260371 8 50 C7.51801325 44.72184999 7.27631227 39.75752734 9.3125 34.8125 C11.90078663 32.10382795 14.57174 29.89259741 17.5 27.5625 C22.01330369 23.95908798 24.93257667 21.41468016 27 16 C27.27815782 13.23794168 27.19566354 10.78820549 27 8 C18.29989657 3.16660921 9.80929858 1.94928696 0 1 C0 0.67 0 0.34 0 0 Z " fill="#92B0FE" transform="translate(51,116)"/>
+          <path d="M0 0 C1.65 1.65 3.3 3.3 5 5 C4.01 5.66 3.02 6.32 2 7 C1.34 7.99 0.68 8.98 0 10 C0 6.7 0 3.4 0 0 Z " fill="#83A6FB" transform="translate(80,124)"/>
+          <path d="M0 0 C0.268125 0.639375 0.53625 1.27875 0.8125 1.9375 C1.88667752 4.33131686 1.88667752 4.33131686 5 5 C5 6.32 5 7.64 5 9 C4.01 8.67 3.02 8.34 2 8 C0.5522078 5.10441559 0 3.25811062 0 0 Z " fill="#7DA6E4" transform="translate(114,158)"/>
+        </svg>
+      }
+      iconType="duotone"
+    >
+  </Card>
+  <Card
       title="Groq"
       href="/latest/integrations/groq"
       icon={
