+++ sdk/python/README.md
@@ -33,7 +33,7 @@ This project adheres to the [Semantic Conventions](https://github.com/open-telem
 | [✅ Ollama](https://docs.openlit.io/latest/integrations/ollama)              | [✅ Pinecone](https://docs.openlit.io/latest/integrations/pinecone)       | [✅ LiteLLM](https://docs.openlit.io/latest/integrations/litellm)         |               |
 | [✅ Anthropic](https://docs.openlit.io/latest/integrations/anthropic)        | [✅ Qdrant](https://docs.openlit.io/latest/integrations/qdrant)           | [✅ LlamaIndex](https://docs.openlit.io/latest/integrations/llama-index)  |               |
 | [✅ GPT4All](https://docs.openlit.io/latest/integrations/gpt4all)            | [✅ Milvus](https://docs.openlit.io/latest/integrations/milvus)           | [✅ Haystack](https://docs.openlit.io/latest/integrations/haystack)       |               |
-| [✅ Cohere](https://docs.openlit.io/latest/integrations/cohere)              |                                                                          | [✅ EmbedChain](https://docs.openlit.io/latest/integrations/embedchain)   |               |                                       
+| [✅ Cohere](https://docs.openlit.io/latest/integrations/cohere)              |                                          | [✅ EmbedChain](https://docs.openlit.io/latest/integrations/embedchain)   |               |                                       
 | [✅ Mistral](https://docs.openlit.io/latest/integrations/mistral)            |                                                                          | [✅ Guardrails](https://docs.openlit.io/latest/integrations/guardrails)   |               |
 | [✅ Azure OpenAI](https://docs.openlit.io/latest/integrations/azure-openai)  |                                                                          |                                       |               |
 | [✅ HuggingFace Transformers](https://docs.openlit.io/latest/integrations/huggingface) |                                                                          |                                        |               |
@@ -43,6 +43,7 @@ This project adheres to the [Semantic Conventions](https://github.com/open-telem
 | [✅ ElevenLabs](https://docs.openlit.io/latest/integrations/elevenlabs)      |                                                                          |                                       |               |
 | [✅ vLLM](https://docs.openlit.io/latest/integrations/vllm)                  |                                                                          |                                       |               |
 | [✅ OLA Krutrim](https://docs.openlit.io/latest/integrations/krutrim)                  |                                                                          |                                       |               |
+| [✅ Google AI Studio](https://docs.openlit.io/latest/integrations/google-ai-studio) |                                                                          |                                       |               |
 
 ## Supported Destinations
 - [✅ OpenTelemetry Collector](https://docs.openlit.io/latest/connections/otelcol)
