+++ sdk/python/src/openlit/__init__.py
@@ -33,6 +33,7 @@ from openlit.instrumentation.ollama import OllamaInstrumentor
 from openlit.instrumentation.gpt4all import GPT4AllInstrumentor
 from openlit.instrumentation.elevenlabs import ElevenLabsInstrumentor
 from openlit.instrumentation.vllm import VLLMInstrumentor
+from openlit.instrumentation.google_ai_studio import GoogleAIStudioInstrumentor
 from openlit.instrumentation.langchain import LangChainInstrumentor
 from openlit.instrumentation.llamaindex import LlamaIndexInstrumentor
 from openlit.instrumentation.haystack import HaystackInstrumentor
@@ -196,6 +197,7 @@ def init(environment="default", application_name="default", tracer=None, otlp_en
         "gpt4all": "gpt4all",
         "elevenlabs": "elevenlabs",
         "vllm": "vllm",
+        "google-ai-studio": "google.generativeai",
         "langchain": "langchain",
         "llama_index": "llama_index",
         "haystack": "haystack",
@@ -273,6 +275,7 @@ def init(environment="default", application_name="default", tracer=None, otlp_en
             "gpt4all": GPT4AllInstrumentor(),
             "elevenlabs": ElevenLabsInstrumentor(),
             "vllm": VLLMInstrumentor(),
+            "google-ai-studio": GoogleAIStudioInstrumentor(),
             "langchain": LangChainInstrumentor(),
             "llama_index": LlamaIndexInstrumentor(),
             "haystack": HaystackInstrumentor(),
