+++ sdk/python/tests/requirements.txt
@@ -16,6 +16,7 @@ botocore>=1.34.93
 google-cloud-aiplatform >= 1.38.1
 groq >= 0.5.0
 elevenlabs >= 1.4.0
+google-generativeai >= 0.7.2
 
 # Vector DBs
 chromadb>=0.4.0
